+++ packages/next/server/base-server.ts
@@ -16,6 +16,7 @@ import type { CacheFs } from '../shared/lib/utils'
 import type { PreviewData } from 'next/types'
 import type { PagesManifest } from '../build/webpack/plugins/pages-manifest-plugin'
 import type { BaseNextRequest, BaseNextResponse } from './base-http'
+import type { PayloadOptions } from './send-payload'
 
 import { join, resolve } from 'path'
 import { parse as parseQs } from 'querystring'
@@ -42,7 +43,7 @@ import * as envConfig from '../shared/lib/runtime-config'
 import { DecodeError, normalizeRepeatedSlashes } from '../shared/lib/utils'
 import { isTargetLikeServerless } from './utils'
 import Router, { replaceBasePath, route } from './router'
-import { PayloadOptions, setRevalidateHeaders } from './send-payload'
+import { setRevalidateHeaders } from './send-payload/revalidate-headers'
 import { IncrementalCache } from './incremental-cache'
 import { execOnce } from '../shared/lib/utils'
 import { isBlockedPage, isBot } from './utils'
